pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.57.1",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "openai>=1.95.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.2",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
]
